+++ glibc/pthread/samples/attr_getdetachstate.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdio.h>
 
-#include <libs/memshow.h>
+#include "libs/memshow.h"
 
 
 void *test_task_fn(void *unused)