+++ glibc/pthread/samples/attr_getschedpolicy.c
@@ -6,7 +6,7 @@
 #include <string.h>
 #include <sched.h>
 
-#include <libs/memshow.h>
+#include "libs/memshow.h"
 
 
 static void print_policy(int policy)
@@ -21,13 +21,13 @@ static void print_policy(int policy)
 void *test_task_fn(void *unused)
 {
 	static int status = 12121;
-	pthread_attr_t attr;
+	pthread_attr_t attr = {};
 	int policy;
 
-	memshow("pthread_attr_t before init ", &attr, sizeof(pthread_attr_t));
+	fhexdump(stdout, "pthread_attr_t before init ", &attr, sizeof(pthread_attr_t));
 
 	pthread_attr_init(&attr);
-	memshow("pthread_attr_t after init  ", &attr, sizeof(pthread_attr_t));
+	fhexdump(stdout, "pthread_attr_t after init  ", &attr, sizeof(pthread_attr_t));
 
 	pthread_attr_getschedpolicy(&attr, &policy);
 	print_policy(policy);
@@ -38,7 +38,7 @@ void *test_task_fn(void *unused)
 	pthread_attr_getschedpolicy(&attr, &policy);
 	print_policy(policy);
 
-	memshow("pthread_attr_t after init  ", &attr, sizeof(pthread_attr_t));
+	fhexdump(stdout, "pthread_attr_t after init  ", &attr, sizeof(pthread_attr_t));
 
 	pthread_attr_destroy(&attr);
 	pthread_exit(&status);