+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 - glibc: test `__GLIBC_USE()`;
 - shell: add some commands from ostools;
 #### Changed
+- pthread: use fhexdump() instead of memshow();
 - Move hexdump() from fakeCUDA to libs/memshow.c;
 - Move vfio into virt/;
 - make: Support NOSUBDIR definition;